backend/requirements.txt
# PASTE THIS IN backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
sqlmodel==0.0.14

# AI/ML
openai==1.3.5
langchain==0.0.346
crewai==0.1.17
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
selenium==4.15.2
webdriver-manager==4.0.1

# PDF Processing
pymupdf==1.23.8
pdfplumber==0.10.3
pillow==10.1.0
pytesseract==0.3.10

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1

# Caching & Queue
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Email
python-dotenv==1.0.0
emails==0.6
jinja2==3.1.2

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0
structlog==23.2.0

# Security
bcrypt==4.1.2
python-magic==0.4.27
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
slowapi==0.1.8